requirement.txt
torch==1.8.1
torchvision==0.9.1+cpu
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
requests-oauthlib==1.3.0
Pillow @ file:///tmp/build/80754af9/pillow_1617383569452/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1618497241363/work
opencv-python==4.5.2.54
matplotlib==3.4.1
matplotlib-inline==0.1.2
Jinja2==3.0.1
jsonify==0.5
Flask==2.0.1